Publish post about Slow is faster
Post on the importance of slowing down in work and life. This post
reflects my recent experiences with Zone2 training and how it taught
me the value of patience and endurance. It emphasizes that sometimes,
to achieve more and perform better, we need to take a slower, more
deliberate approach. Included an illustrative image to enhance the
message.